requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
python-multipart==0.0.9
<<<<<<< HEAD
tensorflow==2.17.1
=======
tensorflow==2.17.0
>>>>>>> 35b665c7f7672268308905183867168de0db2fce
numpy==1.26.3
Pillow==10.2.0
matplotlib==3.8.2
pydantic==2.6.1
